journal/2023: publish "Go test.Data: helper for testing with large output"
One of common Go test pattern is using list of struct with field contains
the input and expected output.
When the expected output is list of records (a slice of struct) with many
fields or text with many lines, filling those fields or writing the multi
lines text become cumbersome and we waste our time by writing the test
data rather than the focus on actual test cases.

The `test.Data` help this by reading the input and expected outputs from
file.
The test function then encode the result to text, for example to JSON,
compare the result by fetching one of expecting output from `test.Data`.