journal/2023: new journal "Using systemd mkosi for development"
In this journal, I will take a notes on my journey learning and using
mkosi as part of integration tests for developing awwan.
The goal is that when I run make on awwan repository, it will create and
run a container using systemd-nspawn(1) with SSH server running, and run
all tests inside the container.