_content: render the metadata "description" on HTML file
The description value, however, is useful for SEO: the description
content value is often what search engines display under the page's
title in search results.
Several browsers, like Firefox and Opera, use this as the default
description of bookmarked pages.
The description should be a short and accurate summary of the page's
content.