journal/2025: new journal "Hyper-V vs VirtualBox"
In this journal we will take a look at the performance of virtual machine
(VM) running on Windows Host using Hyper-V and Oracle VirtualBox.
The guest VM is Arch Linux 64-bit x86-64.
The program that we want to benchmark is an Angular tests suite, an
internal application that I am currently maintenance, running with
Jasmine and Karma using Chromium headless browser.